requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
markovify==0.8.0
mccabe==0.6.1
multidict==4.7.5
pylint==2.4.4
six==1.14.0
Unidecode==1.1.1
uvloop==0.14.0
wavelink==0.6.0
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
